% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19StudyCodeToRun.R
\docType{package}
\name{Covid19StudyCodeToRun}
\alias{Covid19StudyCodeToRun}
\title{DatabaseConnectorAuth}
\description{
This package provides Windows integrated authenication DLL for \code{sqljdbc42}
}
