% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19StudyCodeToRun.R
\name{printCohortSql}
\alias{printCohortSql}
\title{Print out SQL for a cohort}
\usage{
printCohortSql(
  cohortNumber,
  cohortId,
  packageName,
  dbms,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  oracleTempSchema = NULL
)
}
\description{
Print out SQL for a cohort
}
\examples{
printCohortSql(cohortNumber = 1,
               cohortId = 1,
               packageName = "study",
               dbms = "sql server",
               cdmDatabaseSchema = "cdm",
               vocabularyDatabaseSchema = "vocab",
               cohortDatabaseSchema = "cohort_schema",
               cohortTable = "cohort_table",
               oracleTempSchema = NULL)

}
